third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-element-distributed-nodes-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests for a shadow element's getDistributedNodes().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


getDistributedNodes() for "shadow" should return []
PASS distributedNodes.length is expectedDistributedNodes.length

getDistributedNodes() for "host/shadow" should return []
PASS distributedNodes.length is expectedDistributedNodes.length

getDistributedNodes() for "host/shadow" should return []
PASS distributedNodes.length is expectedDistributedNodes.length

getDistributedNodes() for "host/shadow" should return [host-child]
PASS distributedNodes.length is expectedDistributedNodes.length
PASS distributedNodes.item(0) is getNodeInComposedTree(expectedDistributedNodes[0])

getDistributedNodes() for "host/shadow" should return [host-child]
PASS distributedNodes.length is expectedDistributedNodes.length
PASS distributedNodes.item(0) is getNodeInComposedTree(expectedDistributedNodes[0])
PASS successfullyParsed is true

TEST COMPLETE
